requirements.txt
pandas>=1.3.0
matplotlib>=3.4.0
python-docx>=0.8.11
tqdm>=4.62.0
numpy>=1.20.0